Add maximum recursion depth parameter to parser API
Memory usage is already limited by allocating from a user-provided fixed
buffer. But parsing is recursive, so stack usage is only limited by the
input string, which is usually too high. Add a maximum recursion depth
parameter to the parser API, which forces the API user to make the
tradeoff.

This seems to be the most reasonable approach for embedded use. You
could make the parser iterative, and put the stack in the provided
memory (e.g. allocating from the end of the memory region). But most
likely the user would either use recursion on their own, or expect a
shallow JSON AST. In the former case, the user would be forced to limit
stack usage on their own again. This API is simpler, and there is not
much gained from unlimited nesting depth.

RFC 8259 explicitly allows implementations to limit the recursion
depth.