Remove the only use of assert()
The assert is kind of annoying, because it's another case where the libc
might pull in a lot of bloat.

C11 specifies that _Alignof is a power of 2, though the assert() still
would have been nice to detect problems if someone e.g. replaces the
_Alignof use with something pre-C11, or on manual calls to json_alloc().

But we it's unlikely, and it's not worth over accidentally pulling in
stdio. (Still, blame goes to embedded libcs - why do they do something
as stupid as this?)