Change the API and do not use linked lists for arrays/objects
Using linked lists was rather awkward. It wasn't particularly a problem
(usually), because most user code would iterate linearly over arrays and
objects anyway. Random access was probably a rare requirement. But it
was quite awkward. In theory it also wasted some memory.

Linked lists were used because they had a rather nice property: you
didn't need a real memory manager, and could still allocate straight-
forward by bumping a pointer. You could support arbitrary size
and depth of the data structures, limited only by the provided memory
chunk (the depth is also by the CPU stack and the max. depth parameter).

Fortunately, there is a simple trick to represent JSON arrays as C
arrays: put all array items on a "stack", and only move them to the heap
only after the array is fully parsed. This stack is a separate memory
area, that is only used to store array items during parsing. Since
"finished" arrays are moved to the heap, the stack will contain the
array items in a packed way, without anything in-between.

The stack memory is put at the end of the heap, and grows downward. Just
like with traditional stack/heap layouts, this provides maximal
utilization of all memory. In addition, we can easily move array items
from the stack to the heap without increasing memory usage: we can
simply use memmove(). A disadvantage of this arrangement is that the
stack arrays are stored in reversed order, so CPU cycles need to be
wasted for reversing them.

(In theory, reversal and copying could be done at the some time if there
is no overlap, but that seems annoying and probably not worth the
trouble.)

Just making the stack memory completely separate (e.g. by reserving 50%
of the user provided memory or so) would avoid the need for reversal,
but memory utilization would be worse.