Do not use recursion
This seems like a very good property for something that's supposed to
work on embedded. While you can allow recursion, deep stack depths are
always going to be one of the less tested or untested things, and
something to avoid.

Fortunately JSON is simple enough to parse, and changing the recursive
parser to use a shadow stack is relatively simple. It also simplifies
the API by not having to pick a fixed maximum recursion depth. (As long
as you don't use any other recursive functions like json_out_write().)

The alignment static asserts were wrong for the stack case. Add asserts
that get straight to the point. It's easy to see that the alignment
check in json_stack_alloc() is now dead code. (Still leave it, in case
someone tries to remove the asserts.) Types used with json_alloc() on
the other hand can serve allocations with smaller size/alignment,
because it is allowed to add padding after them. I hope the code
comments are clear enough.

If we couldn't rely on the MAX_ALIGN magic, then json_stack_alloc()
would have to perform alignment, and we'd have to deal with the padding
this would possibly add, for example by keeping track of an explicit
items data start pointer and an explicit before-curlist pointer per
curlist stack frame. For now, it seems nicer to rely on MAX_ALIGN.