Add a way to output JSON to a callback
This is useful if you can "stream" out the data in a blocking manner, or
if you have some other buffer somewhere, that requires calling a
function to fill it.

In theory, this isn't quite ideal. The callback overhead might be much
larger than writing to some memory for small strings. An ideal
implementation would probably be a hybrid temporary buffer + callback to
"flush" it (similar to how stdio usually implements FILE). But I didn't
want to overdo it. Requiring such a temporary buffer would either
complicate the API (having to provide it), or bloat json_out (if we just
make it a fixed inline buffer).

The additional stack buffer might raise the stack size requirement
(issue for embedded software), but I don't think it's crazy.