Change when list structs are allocated
Allocate the list structs (json_object, json_array) before instead of
after parsing the list contents. Replace the tricky stack pointer
comparison (that checks whether a list is empty) with something more
explicit.

This is preparation for adding malloc() support, for which this code
structure is better.

Also change a bogus "return NULL" to "return false" (C usually defines
both NULL and false to 0, so this could not cause a warning).