Improve handling of memory alignment
Some improvements; this probably also fixes some corner cases, which
could lead to unaligned accesses (but probably not on most platforms
assuming unmodified code, so this is just for robustness). It also can
help reduce memory requirements very slightly.

Use an union to let the C compiler compute the required alignment,
instead of picking the likely worst case type for it. This doesn't
loosen any alignment requirements, but allows us to drop some of the
static_assert()s.

Use a better trick for curlist_alloc, which in theory saves some memory.
For example, if on a 32 bit platform alignof(json_tok)==8, and its size
is 16, then curlist_alloc can be 8 bytes instead of 16 bytes.

Use % in the static_assert()s instead of bit-twiddling, it's slightly
more readable.

Adjust comments and the README. The suggestion to set MAX_ALIGN to 8 has
been wrong for a while (although it would work on most platforms).

I considered simplifying this and abandoning the requirement on the
tight alignment requirements by doing allocation in a more "straight"
way. But that would have meant adding an additional pointer to curlist,
and extra calls to align the stack explicitly to keep the way how list
parsing is done working. It didn't seem like a huge win after all,
although the concerns on theoretical portability and extensibility would
have disappeared.