Limit maximum parser depth to 64 by default
Before this, the parse depth was unlimited (restricted by available
memory). This commit limits it by default.

Do this because an unlimited depth would be such an easy way to DoS
anyone using it. It's unlikely anyone really wants unlimited depth,
and the chosen value 64 ought to be enough for everyone. If it isn't
you can still adjust the level by setting json_parse_opts.depth. This
is only a slight inconvenience for those who need it, and it's better
than being DoSable by default (if the API user happens to use a
recursive function on the resulting parse tree later).

Technically, this is an API change (even if unlikely to cause problems).
I will start to list such changes in the README.