Optional malloc() support
Currently, the parser uses a single fixed size memory region as work
memory and to allocate the result json_tok tree. Sometimes, such as
in non-embedded settings, you may prefer a malloc()'ed result.

This adds an optional mrealloc callback. This works mostly like
realloc(), and will be used to allocate the json_tok result.
json_helpers_malloc.c/.h uses this callback and provides a sane
malloc-based API to the parser. It also offers a large number of
accessors.

This is a bit wild, because of the tricks the code uses to avoid
recursion all while working with a single memory block and
outputting a tree structure with variable length arrays in its nodes.
And now there needs to be a code path which uses dynamic memory
allocation anyway. But conceptually, it's somehow still pretty straight
forward, at least considering it's sharing the crazy static memory
without recursion code path.

So I guess it's not too bad, even if it increases complexity.

While the parser always avoids recursion, the no-recursion requirement
is dropped for the malloc API. If you got malloc, you probably got a
huge stack as well, so I don't see this as a problem.

The test program becomes even more convoluted and messy, but that's
what it's there for.