Support trailing ',' as extension
Generally useful. Consider git for example: you can add a new entry at
the end of a JSON object or array, without having to touch the previous
line, which reduces the noise in the diff.

Standard JSON doesn't allow this, so enable_extensions must be set to
true, or the parser rejects it.