Copy string values instead of entire input
There's is a normal and a "destructive" parsing mode. Destructive
parsing is basically the default: it modifies the JSON input text to
process string tokens and to turn them into 0-terminated C strings. The
normal parsing mode simply copied the entire input to avoid modifying
the input.

This commit changes it to copy only the strings, instead of the entire
input. The disadvantage is that it requires 2 passes now, which might
make it slower. The destructive parsing mode still works the same. (And
it's useful, minimizes memory usage.)

Allocating the strings is done better without alignment, so add a
json_alloc_align() with an alignment value. json_alloc() still does the
old thing (could change all callers to pass the actual alignment, but in
practice it's always the same value anyway, and some callers make it
annoying).