Change another implementation detail to reduce shadow stack size
The recent alignment handling change added another pointer field to
struct curlist. This can be "easily" avoided by using some trickery
(which is better than the old trickery). Also, this makes curlist.tok
unnecessary for the parsing logic itself, which is a preparation for the
pull API addition. The parsing logic only needs the token type, which
can be stored separately as 1-bit bool, and which makes the code
actually slightly simpler. (prev_offset might be too small in
theoretical scenarios, could solve this by dividing by the type's
alignment and shifting away the unused LSBs. But this seems far too
involved for a problem that isn't real so I'm just ignoring it.)