Add a pull API
The pull API doesn't build a token tree and leaves this to the user.
This is good if memory becomes a problem. The token memory usually uses
up much more memory than the input text. With this API, only memory for
the shadow stack has to be provided.

The disadvantage is that it's much harder to use.

With the previous changes, this requires minimal changes. The code
duplication is fairly minimal as well. But of course the truth is that
there are many weird code paths now because of the pull API and the
malloc support and the constant C stack usage change, and the original
simplicity is all lost.

Maybe the pull code should be separated, and the AST parser should use
it as library. But the pull code needs to track the stack as well to
report syntax errors. A particularly simpler and resource-constrained
user might be fine with incorrect and mismatched array/object nesting,
but I don't want to accept such input. (But there are other arguments,
for example the pull code could keep track of the nesting with a simple
bitmap, or do multiple passes over the input.)

The actually duplicated parts are parse_ast() and json_pull_next(),
it's not much, basically they're simply using some sort of internal pull
API.

TODO: do I really want this?