Don't allow whitespace between two surrogate escape pairs
RFC 8259 clearly calls it "a 12-character sequence". We skipped the
whitespace and ignored it, instead of raising an error.

Fix this indirectly by making skip_str() not skip whitespace. (The
surrogate parsing code used skip_str() to check and skip the next escape
header.) Add skip_ws() to all callers which actually need to skip
whitespace.